Delete track when account is deleted

+++ src/routes/account.js
@@ -1,9 +1,11 @@
+/* eslint-disable consistent-return */
 /* eslint-disable no-underscore-dangle */
 const express = require('express');
 const mongoose = require('mongoose');
 const requireAuth = require('../middlewares/requireAuth');
 
 const User = mongoose.model('User');
+const Track = mongoose.model('Track');
 
 const router = express.Router();
 router.use(requireAuth);
@@ -21,6 +23,12 @@ router.post('/edit-account', async (req, res) => {
 
 router.delete('/delete-account', async (req, res) => {
   try {
+    const trails = await Track.find({ userId: req.user._id });
+    const trailIds = trails.map(({ id }) => id);
+    const track = await Track.deleteMany({ _id: { $in: trailIds } });
+    if (!track.acknowledged) {
+      return res.status(400).send({ message: 'Failed to delete this account, Please try again later' });
+    }
     const user = await User.findOneAndDelete({ _id: req.user._id });
     if (user) {
       res.status(200).send({
